Changed according to new swagger
