+++ Shuttle-back/src/main/java/com/shuttle/note/dto/NoteDTO.java
@@ -1,11 +1,11 @@
 package com.shuttle.note.dto;
 
-import java.util.Date;
+import java.time.ZonedDateTime;
 
 public class NoteDTO {
 
 	private long id;
-	private Date date;
+	private ZonedDateTime date;
 	private String message;
 	public long getId() {
 		return id;
@@ -13,17 +13,18 @@ public class NoteDTO {
 	public void setId(long id) {
 		this.id = id;
 	}
-	public Date getDate() {
-		return date;
-	}
-	public void setDate(Date date) {
-		this.date = date;
-	}
+
 	public String getMessage() {
 		return message;
 	}
 	public void setMessage(String message) {
 		this.message = message;
+	}
+	public ZonedDateTime getDate() {
+		return date;
+	}
+	public void setDate(ZonedDateTime date) {
+		this.date = date;
 	}
