--- Shuttle-back/src/main/java/com/shuttle/ride/dto/ReadRideDTO.java
@@ -1,17 +1,0 @@
-package com.shuttle.ride.dto;
-
-public class ReadRideDTO extends BaseRideDTO{
-
-	private String status;
-	private RejectionDTO rejection;
-
-	public String getStatus() {
-		return status;
-	}
-	public void setStatus(String status) {
-		this.status = status;
-	}
-	
-	
-
-}
