+++ Shuttle-back/src/main/java/com/shuttle/user/dto/UserDTO.java
@@ -2,17 +2,13 @@ package com.shuttle.user.dto;
 
 public class UserDTO {
 	private long id;
+	private String name;
 	private String surname;
 	private String profilePicture;
 	private String telephoneNumber;
 	private String email;
 	private String address;
-	public long getId() {
-		return id;
-	}
-	public void setId(long id) {
-		this.id = id;
-	}
+
 	public String getSurname() {
 		return surname;
 	}
@@ -42,6 +38,18 @@ public class UserDTO {
 	}
 	public void setAddress(String address) {
 		this.address = address;
+	}
+	public long getId() {
+		return id;
+	}
+	public void setId(long id) {
+		this.id = id;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
 	}
